Move document signature forward to esign function
