+++ src/app/GraphQL/Mutations/DocumentSignatureMutator.php
@@ -3,12 +3,15 @@
 namespace App\GraphQL\Mutations;
 
 use App\Enums\DocumentSignatureSentNotificationTypeEnum;
+use App\Enums\PeopleGroupTypeEnum;
 use App\Enums\SignatureStatusTypeEnum;
 use App\Http\Traits\SendNotificationTrait;
 use App\Http\Traits\SignatureTrait;
 use App\Exceptions\CustomException;
 use App\Models\DocumentSignature;
+use App\Models\DocumentSignatureForward;
 use App\Models\DocumentSignatureSent;
+use App\Models\People;
 use Carbon\Carbon;
 use Illuminate\Http\Response;
 use Illuminate\Support\Arr;
@@ -144,6 +147,7 @@ class DocumentSignatureMutator
             throw new CustomException('Webhook failed', json_decode($response));
         } else {
             $data = $this->updateDocumentSentStatus($data, $newFileName, $verifyCode);
+            $this->forward($data);
         }
 
         Storage::disk('local')->delete($newFileName);
@@ -222,12 +226,128 @@ class DocumentSignatureMutator
     protected function addFooterDocument($data, $newFileName, $verifyCode)
     {
         $addFooter = Http::post(config('sikd.add_footer_url'), [
-            'pdf' => $data->documentSignature->url,
+            'pdf' => 'https://devsimanis.jabarprov.go.id/api/mobile/api/v1/draft/218180322035035',
             'qrcode' => config('sikd.url') . '/FilesUploaded/ttd/sudah_ttd/' . $newFileName,
             'category' => $data->documentSignature->documentSignatureType->document_paper_type,
             'code' => $verifyCode
         ]);
 
         return $addFooter;
+    }
+
+    protected function forward($documentSignatureSent)
+    {
+        $nextDocument = DocumentSignatureSent::where('id', $documentSignatureSent->id)
+                                            ->where('urutan', $documentSignatureSent->urutan + 1)
+                                            ->first();
+
+        if (!$nextDocument) {
+            $documentSignatureForwardIds = $this->doForward($documentSignatureSent);
+
+            if (!$documentSignatureForwardIds) {
+                throw new CustomException(
+                    'Forward document failed',
+                    'Return ids is missing. Please try again.'
+                );
+            }
+
+            $this->doSendForwardNotification($documentSignatureSent->id, $documentSignatureSent->receiver->PeopleName);
+        }
+
+        return $documentSignatureForwardIds;
+    }
+
+    /**
+     * doSendForwardNotification
+     *
+     * @param  string $id
+     * @param  string $name
+     * @return void
+     */
+    protected function doSendForwardNotification($id, $name)
+    {
+        $messageAttribute = [
+            'notification' => [
+                'title' => 'TTE Naskah',
+                'body' => 'Naskah Anda telah di tandocumentSignatureSentngani oleh ' . $name . '. Klik disini untuk lihat naskah!',
+            ],
+            'data' => [
+                'documentSignatureSentId' => $id,
+                'target' => DocumentSignatureSentNotificationTypeEnum::SENDER()
+            ]
+        ];
+
+        $this->setupDocumentSignatureSentNotification($messageAttribute);
+    }
+
+    /**
+     * doForward
+     *
+     * @param  object $documentSignatureSentId
+     * @param  string $sender
+     * @param  mixed $args
+     * @return array
+     */
+    public function doForward($documentSignatureSent)
+    {
+        $ids = array();
+        $receiver = $this->forwardReceiver($documentSignatureSent);
+
+        if ($receiver != null) {
+            foreach ($receiver as $key => $receiver) {
+                $key++;
+                $documentSignatureForward = DocumentSignatureForward::create([
+                    'ttd_id' => $documentSignatureSent->ttd_id,
+                    'catatan' => '',
+                    'tgl' => Carbon::now(),
+                    'PeopleID' => $documentSignatureSent->PeopleIDTujuan,
+                    'PeopleIDTujuan' => $receiver,
+                    'urutan' => $key,
+                    'status' => SignatureStatusTypeEnum::WAITING()->value,
+                ]);
+
+                array_push($ids, $documentSignatureForward);
+            }
+
+            return $ids;
+        }
+
+        return false;
+    }
+
+    /**
+     * forwardReceiver
+     *
+     * @param  mixed $type
+     * @return mixed
+     */
+    public function forwardReceiver($documentSignatureSent)
+    {
+        $type = optional($documentSignatureSent->documentSignature->documentSignatureType)->document_forward_target;
+        switch ($type) {
+            case 'UK':
+            case 'TU':
+                if ($type == 'UK') {
+                    $peopleGroupType = PeopleGroupTypeEnum::UK()->value;
+                    $whereField = 'GRoleId';
+                    $whereParams = auth()->user()->role->GRoleId;
+                }
+                if ($type == 'TU') {
+                    $peopleGroupType = PeopleGroupTypeEnum::TU()->value;
+                    $whereField = 'Code_Tu';
+                    $whereParams = auth()->user()->role->Code_Tu;
+                }
+                $receiver = People::whereHas('role', function ($role) use ($whereField, $whereParams) {
+                    $role->where('RoleCode', auth()->user()->role->RoleCode);
+                    $role->where($whereField, $whereParams);
+                })->where('GroupId', $peopleGroupType)->pluck('PeopleId');
+                break;
+
+            default:
+                $receiver = null;
+                break;
+        }
+
+        return $receiver;
     }
 }
